mmngr_drv: mmngrbuf: Take care of any migration of scatterlist
MMNGRBUF should take care of any migration of scatterlist for
all (shared) users of the DMA buffer. Although we only care about
DMA addresses at the moment, this will benefit the DRM system that
uses the buffer allocated by MMNGR directly.

Initialize the scatterlist entry so that it can be happily traversed.

Signed-off-by: Hai Nguyen Pham <hai.pham.ud@renesas.com>